Move SubImage into separate module
